Backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
crewai
crewai-tools
google-generativeai
litellm
groq
yt-dlp
opencv-python
torch
torchvision
transformers
faiss-cpu
Pillow
numpy
requests
ImageHash
